--- app/src/main/java/com/example/quizapp/ParcelableQuiz.kt
@@ -1,30 +1,0 @@
-package com.example.quizapp
-
-import android.os.Parcel
-import android.os.Parcelable
-
-/*
-class ParcelableQuiz() : Parcelable, Serializable {
-    constructor(parcel: Parcel) : this() {
-    }
-
-    override fun writeToParcel(parcel: Parcel, flags: Int) {
-
-    }
-
-    override fun describeContents(): Int {
-        return 0
-    }
-
-    companion object CREATOR : Parcelable.Creator<ParcelableQuiz> {
-        override fun createFromParcel(parcel: Parcel): ParcelableQuiz {
-            return ParcelableQuiz(parcel)
-        }
-
-        override fun newArray(size: Int): Array<ParcelableQuiz?> {
-            return arrayOfNulls(size)
-        }
-    }
-}
-
- */
