+++ app/src/main/java/com/example/quizapp/QuizModifier.kt
@@ -6,7 +6,6 @@ import android.graphics.Color
 import android.net.Uri
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.util.Log
 import android.view.Menu
 import android.view.MenuItem
 import android.widget.ArrayAdapter
@@ -14,6 +13,7 @@ import android.widget.Button
 import android.widget.Spinner
 import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.widget.Toolbar
+import com.example.quizapp.localclasses.ThemeConverter
 
 class QuizModifier : AppCompatActivity() {
