+++ app/src/main/java/com/example/quizapp/SavedQuiz.kt
@@ -6,13 +6,12 @@ import android.content.Intent
 import android.graphics.Color
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.util.Log
 import android.view.*
-import android.view.View.inflate
 import android.widget.*
 import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.widget.Toolbar
-import com.example.quizapp.test.TestDatabase
+import com.example.quizapp.database.QuizOpener
+import com.example.quizapp.localclasses.FullQuiz
 import java.io.Serializable
 
 class SavedQuiz : AppCompatActivity() {
