--- app/src/main/java/com/example/quizapp/StoreData.kt
@@ -1,5 +1,0 @@
-package com.example.quizapp
-
-interface StoreData {
-    fun storedData(curQuiz: FullQuiz)
-}
