+++ app/src/main/java/com/example/quizapp/api/StoreData.kt
@@ -1,0 +1,7 @@
+package com.example.quizapp.api
+
+import com.example.quizapp.localclasses.FullQuiz
+
+interface StoreData {
+    fun storedData(curQuiz: FullQuiz)
+}
