+++ app/src/main/java/com/example/quizapp/database/QuestionOpener.kt
@@ -1,11 +1,11 @@
-package com.example.quizapp
+package com.example.quizapp.database
 
 import android.content.ContentValues
 import android.content.Context
 import android.database.sqlite.SQLiteDatabase
 import android.database.sqlite.SQLiteOpenHelper
 import android.provider.BaseColumns
-import android.util.Log
+import com.example.quizapp.localclasses.QuizQuestion
 
 class QuestionOpener(context: Context): SQLiteOpenHelper(context, DATABASE_NAME, null, DATABASE_VERSION) {
     override fun onCreate(db: SQLiteDatabase) {
