+++ app/src/main/java/com/example/quizapp/localclasses/FullQuiz.kt
@@ -1,4 +1,4 @@
-package com.example.quizapp
+package com.example.quizapp.localclasses
 
 import android.util.Log
 import java.io.Serializable
